+++ src/test/suite/menu/whichkeyMenu.test.ts
@@ -1,6 +1,6 @@
 import { commands, Disposable } from "vscode";
 import { CommandRelay } from "../../../commandRelay";
-import { ActionType } from "../../../config/bindingItem";
+import { ActionType, DisplayOption } from "../../../config/bindingItem";
 import { WhichKeyMenuConfig } from "../../../config/menuConfig";
 import { showWhichKeyMenu } from "../../../menu/whichKeyMenu";
 import { StatusBar } from "../../../statusBar";
@@ -52,5 +52,46 @@ suite("WhichKeyMenu", function () {
 		showWhichKeyMenu(statusBar, cmdRelay, undefined, config);
 		cmdRelay.triggerKey("m");
 		cmdRelay.triggerKey("x");
+	});
+
+	test("can hide item with `display: hidden`", function (done) {
+		const statusBar = new StatusBar();
+		const cmdRelay = new CommandRelay();
+		const config: WhichKeyMenuConfig = {
+			delay: 0,
+			showIcons: false,
+			title: "Test",
+			bindings: [
+				{
+					key: "a",
+					name: "Should be hidden",
+					type: ActionType.Command,
+					display: DisplayOption.Hidden,
+					command: "whichkey.hiddenCommand"
+				},
+				{
+					key: "x",
+					name: "test command",
+					type: ActionType.Command,
+					command: "whichkey.testCommand"
+				}
+			]
+		};
+		disposables = [
+			statusBar,
+			cmdRelay,
+			commands.registerCommand("whichkey.testCommand", () => {
+				done();
+			}),
+			commands.registerCommand("whichkey.hiddenCommand", () => {
+				done("The item should be hidden");
+			})
+		];
+
+		showWhichKeyMenu(statusBar, cmdRelay, undefined, config);
+		// Wait until the UI is shown
+		setTimeout(() => {
+			commands.executeCommand("workbench.action.acceptSelectedQuickOpenItem");
+		}, 100);
 	});
 });